refactor(speech to text): Add manual post-gen changes to file using monkeypatching
